task3/requirements.txt
mysql-connector-python=8.0.23